+++ core/src/test/scala/cz/cvut/fit/prl/scalaimplicit/core/extraction/NewSchemaTest.scala
@@ -142,35 +142,29 @@ class NewSchemaTest extends SemanticdbTest {
               isImplicit = true,
               signature = Some(
                 Signature(
-                  typeParams = Seq(),
+                  typeParameters = Seq(),
                   parameterLists = Seq(
-                    DeclaredParameterList(
-                      isImplicit = false,
-                      params = Seq(
-                        DeclaredParameter(
-                          name = "a",
-                          tipe = Type("nested.A")
-                        ))
-                    ),
-                    DeclaredParameterList(
-                      isImplicit = true,
-                      params = Seq(
-                        DeclaredParameter(
-                          name = "b",
-                          tipe = Type(name = "scala.Function1",
-                                      parameters =
-                                        Seq(Type("scala.Function1.T1"),
-                                            Type("scala.Function1.R")))
-                        ),
-                        DeclaredParameter(
-                          name = "c",
-                          tipe = Type(name = "scala.Function1",
-                                      parameters =
-                                        Seq(Type("scala.Function1.T1"),
-                                            Type("scala.Function1.R")))
-                        )
+                    DeclaredParameterList(isImplicit = false, parameters = Seq(
+                      DeclaredParameter(
+                        name = "a",
+                        parameterType = Type("nested.A")
+                      ))),
+                    DeclaredParameterList(isImplicit = true, parameters = Seq(
+                      DeclaredParameter(
+                        name = "b",
+                        parameterType = Type(name = "scala.Function1",
+                          typeParameters =
+                            Seq(Type("scala.Function1.T1"),
+                              Type("scala.Function1.R")))
+                      ),
+                      DeclaredParameter(
+                        name = "c",
+                        parameterType = Type(name = "scala.Function1",
+                          typeParameters =
+                            Seq(Type("scala.Function1.T1"),
+                              Type("scala.Function1.R")))
                       )
-                    )
+                    ))
                   ),
                   returnType = Some(Type("nested.C"))
                 )
@@ -193,16 +187,13 @@ class NewSchemaTest extends SemanticdbTest {
                   isImplicit = true,
                   signature = Some(
                     Signature(
-                      typeParams = Seq(),
+                      typeParameters = Seq(),
                       parameterLists = Seq(
-                        DeclaredParameterList(
-                          isImplicit = false,
-                          params = Seq(
-                            DeclaredParameter(
-                              name = "a",
-                              tipe = Type("nested.A")
-                            ))
-                        )),
+                        DeclaredParameterList(isImplicit = false, parameters = Seq(
+                          DeclaredParameter(
+                            name = "a",
+                            parameterType = Type("nested.A")
+                          )))),
                       returnType = Some(Type("nested.B"))
                     )
                   )
@@ -224,16 +215,13 @@ class NewSchemaTest extends SemanticdbTest {
                   isImplicit = true,
                   signature = Some(
                     Signature(
-                      typeParams = Seq(),
+                      typeParameters = Seq(),
                       parameterLists = Seq(
-                        DeclaredParameterList(
-                          isImplicit = false,
-                          params = Seq(
-                            DeclaredParameter(
-                              name = "b",
-                              tipe = Type("nested.B")
-                            ))
-                        )),
+                        DeclaredParameterList(isImplicit = false, parameters = Seq(
+                          DeclaredParameter(
+                            name = "b",
+                            parameterType = Type("nested.B")
+                          )))),
                       returnType = Some(Type("nested.C"))
                     )
                   )
@@ -289,27 +277,21 @@ class NewSchemaTest extends SemanticdbTest {
           isImplicit = true,
           signature = Some(
             Signature(
-              typeParams = Seq(Type("classConv.Hello.T")),
+              typeParameters = Seq(Type("classConv.Hello.T")),
               parameterLists = Seq(
-                DeclaredParameterList(
-                  isImplicit = false,
-                  params =
-                    Seq(DeclaredParameter("s", Type("classConv.Hello.T")))
-                ),
-                DeclaredParameterList(
-                  isImplicit = true,
-                  params = Seq(
-                    DeclaredParameter(
-                      name = "evidence$1",
-                      tipe = Type(
-                        name = "classConv.Writer",
-                        parameters = Seq(Type("classConv.Writer.A"))
-                      )
-                    ))
-                )
+                DeclaredParameterList(isImplicit = false, parameters =
+                  Seq(DeclaredParameter("s", Type("classConv.Hello.T")))),
+                DeclaredParameterList(isImplicit = true, parameters = Seq(
+                  DeclaredParameter(
+                    name = "evidence$1",
+                    parameterType = Type(
+                      name = "classConv.Writer",
+                      typeParameters = Seq(Type("classConv.Writer.A"))
+                    )
+                  )))
               ),
               returnType =
-                Some(Type("classConv.Hello[T]", parameters = Seq(Type("T"))))
+                Some(Type("classConv.Hello[T]", typeParameters = Seq(Type("T"))))
             )),
           parents = Seq()
         ),
@@ -351,7 +333,7 @@ class NewSchemaTest extends SemanticdbTest {
                     isImplicit = false,
                     signature = Some(
                       Signature(
-                        typeParams = Seq(Type("classConv.Writer.A")),
+                        typeParameters = Seq(Type("classConv.Writer.A")),
                         returnType = Some(Type("<notype>"))
                       ))
                   ),
