+++ queries/src/main/scala/cz/cvut/fit/prl/scalaimplicit/queries/package.scala
@@ -1,0 +1,7 @@
+package cz.cvut.fit.prl.scalaimplicit
+
+package object queries {
+
+  object implicits extends ImplicitMatchers
+
+}
